backend/requirements.txt
fastapi
fastapi[standard]
uvicorn
pydantic
firebase-admin
twilio
python-dotenv
